+++ modules/hosts/valkyrie/default.nix
@@ -82,7 +82,10 @@ in {
 
       system = {
         scheme = "secure";
-        fs.scheme = "advanced";
+        fs = {
+          scheme = "advanced";
+          disk = "/dev/disk/by-id/nvme-eui.00000000000000000026b7686ab56005";
+        };
 
         # Kernel
         kernel = "zen";